refactor(domain): extract shared agent kind - add the agent_kind domain module - move stable identifier parsing and formatting - update comparison imports
